frame: add one param if a non-static method is called
from the spec:
For example, the method descriptor for the method

    Object mymethod(int i, double d, Thread t)

is

	(IDLjava/lang/Thread;)Ljava/lang/Object;

The method descriptor for `mymethod' is the same whether `mymethod' is a
class or an instance method. Although an instance method is passed this,
a reference to the current class instance, in addition to its intended
parameters, that fact is not reflected in the method descriptor. (A
reference to this is not passed to a class method.) The reference to
this is passed implicitly by the method invocation instructions of the
Java virtual machine used to invoke instance methods.